Do a big rename of editor related components
I don't want to use orthogonal "views" in my routing so I want to get
rid of that veribiage, and the names could've been better to start with.
