+++ client/components/editor/editor.module.css.d.ts
@@ -1,16 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'active': string;
-  'editor': string;
-  'error': string;
-  'error-marker': string;
-  'error-overlay': string;
-  'errorMarker': string;
-  'errorOverlay': string;
-  'tab': string;
-  'tab-bar': string;
-  'tabBar': string;
+  'layout': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
