+++ client/components/program-editor/program-editor.module.css
@@ -1,0 +1,58 @@
+.editor {
+  position: relative;
+  color: var(--text-color);
+  background-color: var(--background-color-light);
+  display: grid;
+  grid-template-rows: min-content auto;
+}
+
+.error-marker {
+  position: absolute;
+  z-index: 3;
+  border-radius: 0;
+  border-bottom: 1px dotted var(--error-color);
+}
+
+.tab-bar {
+  display: flex;
+  border-bottom: 1px solid var(--border-color-dark);
+}
+
+.tab {
+  height: 30px;
+  padding: 0 10px;
+  font-size: 1.1rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+  cursor: pointer;
+  user-select: none;
+}
+.tab.active {
+  border-bottom: 2px solid var(--accent-color-orange);
+  background-color: var(--background-color-highlighted);
+}
+.tab.error {
+  color: var(--error-color);
+}
+.tab.error::after {
+  content: '!';
+  font-weight: bold;
+}
+
+.error-overlay {
+  --inset: 10px;
+  position: absolute;
+  bottom: var(--inset);
+  left: calc(48px + var(--inset));
+  right: var(--inset);
+  z-index: 5;
+  padding: 8px;
+  color: var(--text-color);
+  background-color: var(--background-color-light);
+  border: 1px solid var(--error-color);
+  font-family: var(--monospace-font-family);
+  font-size: 1.2rem;
+  font-weight: bold;
+}
