--- client/components/views/editor-view/editor-view.module.css
@@ -1,10 +1,0 @@
-.layout {
-  height: 100vh;
-  display: grid;
-  grid-template-areas:
-    "header header"
-    "editor scene"
-  ;
-  grid-template-rows: min-content auto;
-  grid-template-columns: 1fr 1fr;
-}
