--- client/components/views/editor-view/editor-view.module.css.d.ts
@@ -1,7 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'layout': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
